requirements.txt
matplotlib>=3.6.2
numpy>=1.23.4
pandas>=1.5.1
scikit-learn>=1.1.3
scikit-dimension>=0.3.2
tqdm>=4.64.1
torch>=1.13.0
skorch>=0.12.1
torchvision>=0.14.0
torchmetrics>=0.10.3
torchsummary>=1.5.1
hydra-core>=1.3.0
numba>=0.56.4
